PAGINA_WEB/codigo_web/requirements.txt
Flask
gunicorn
paho-mqtt
qrcode[pil]